Soon api

+++ src/database/migrations/20230206015128-create-veiculo.js
@@ -45,7 +45,7 @@ module.exports = {
         allowNull: false, 
       },
       distancia_total: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(15, 2),
         allowNull: false, 
       },
       duracao_total: {
